requirements.txt
google-cloud-aiplatform
langchain-google-vertexai
langchain
flask
gunicorn
